vuedjango/requirement.txt
asgiref==3.5.0
beautifulsoup4==4.11.1
Django==4.0.4
django-bootstrap-v5==1.0.11
django-cors-headers==3.13.0
django-markdown-deux==1.0.5
django-sslserver==0.22
django-webpack-loader==1.6.0
djangorestframework==3.13.1
markdown2==2.4.3
mysqlclient==2.1.1
npm==0.1.1
optional-django==0.1.0
pytz==2022.1
soupsieve==2.3.2.post1
sqlparse==0.4.2
tzdata==2022.1
